Apply `@final` to all payloads
Pyright uses this to more confidently narrow down TypedDict since a missing field doesn't mean that it can't be supplied by a subclass (unless it is marked as final).

Mypy does not support this at all though. Supporting Mypy would come with a significant disadvantage for other users of Pyright so it is not a priority right now for the sake of progressing the library.